+++ CHANGELOG.md
@@ -12,10 +12,15 @@ Unless otherwise specified, any version comparison below is the comparison of th
 
 <small>[Compare with 0.10.5](https://github.com/AllayMC/Allay/compare/0.10.5...HEAD)</small>
 
+### Added
+
+- (API) Added a overload to method `WorldViewer.viewPlayerEmote` which accepts a boolean `silence` parameter.
+
 ### Fixed
 
 - Fixed a bug where skin can't be set for fake player.
 - Fixed duplicate "Client disconnected" messages during server shutdown caused by players not being removed from the dimension's player list when disconnecting.
+- Fixed method `WorldViewer.viewPlayerEmote` not working.
 
 # 0.10.5 (API 0.22.0) - 2026/1/15